+++ app/src/main/java/com/ar/openClimbAR/utils/MapViewWidget.java
@@ -183,7 +183,7 @@ public class MapViewWidget {
         Drawable nodeIcon = osmMap.getContext().getResources().getDrawable(R.drawable.ic_topo_small);
         nodeIcon.mutate(); //allow different effects for each marker.
 
-        float remapGradeScale = ArUtils.remapScale(0f,
+        float remapGradeScale = (float)ArUtils.remapScale(0f,
                 GradeConverter.getConverter().maxGrades,
                 1f,
                 0f,
@@ -227,7 +227,7 @@ public class MapViewWidget {
             doAutoCenter = true;
         }
 
-        obsLocationMarker.setRotation(Globals.observer.degAzimuth);
+        obsLocationMarker.setRotation((float)Globals.observer.degAzimuth);
         obsLocationMarker.getPosition().setCoords(Globals.observer.decimalLatitude, Globals.observer.decimalLongitude);
         obsLocationMarker.getPosition().setAltitude(Globals.observer.elevationMeters);
